refactor(passthrough): hoist alt-screen helpers into passthrough_replay
ALT_SCREEN_ENTER/EXIT + contains_alt_screen_enter/exit +
update_host_alt_screen lived in pane_io.rs as private helpers, but
they describe a concern that belongs to the passthrough machinery
(detecting / mirroring inner-PTY alt-screen state). Moving them to
passthrough_replay.rs as `pub(crate)` so the replay-log layer can
reuse the same detectors for proactive snapshot refresh on
alt-screen exit (next commit).

Pure refactor, no behavior change. pane_io.rs now imports them at
the top; the existing tests submodule's reference to
update_host_alt_screen routes through the new home.
